agent/requirements.txt
opencv-python-headless==4.9.0.80
httpx==0.27.0
sqlalchemy==2.0.29
pandas==2.2.1
openpyxl==3.1.2
pydantic==2.7.0
python-dotenv==1.0.1
tenacity==8.2.3
Pillow==10.3.0
numpy==1.26.4
watchdog==4.0.0
pytesseract==0.3.10
